Fixed edit button listeners
Added prototype class to show if it was the first time the update edit button was clicked or if it was the first time the update tab was opened. Supposedly fixed listener stacking for save and cancel button. Before, it was creating a new click event listener for the save button every time the edit button was clicked. This was causing unnecessary writes to the database.