Made encrypt/decrypt master pass function efficient
Tested to make sure requests to firebase were being made the correct number of times.

Also, before this commit, everytime we decrypted/encrypted something with the master password hash, we were sending a read to Firestore to get the master password hash. Now, we store that hash in a variable so we don't keep on having to make that read.